Fix gcc warnings

+++ apps/atlasRobot/Controller.cpp
@@ -147,7 +147,7 @@ void Controller::changeStateMachine(const string& _name, double _currentTime)
 }
 
 //==============================================================================
-void Controller::changeStateMachine(int _idx, double _currentTime)
+void Controller::changeStateMachine(size_t _idx, double _currentTime)
 {
   assert(0 <= _idx && _idx <= mStateMachines.size()
          && "Invalid index of StateMachine.");
@@ -211,7 +211,7 @@ void Controller::printDebugInfo() const
             << " NUM DOF   : " << mAtlasRobot->getDof() << std::endl
             << " NUM JOINTS: " << mAtlasRobot->getNumBodyNodes() << std::endl;
 
-  for(int i = 0; i < mAtlasRobot->getNumBodyNodes(); ++i)
+  for(size_t i = 0; i < mAtlasRobot->getNumBodyNodes(); ++i)
   {
     Joint* joint = mAtlasRobot->getJoint(i);
     BodyNode* body = mAtlasRobot->getBodyNode(i);
@@ -814,12 +814,12 @@ StateMachine* Controller::_createRunningStateMachine()
 //==============================================================================
 void Controller::_setJointDamping()
 {
-  for (int i = 1; i < mAtlasRobot->getNumBodyNodes(); ++i)
+  for (size_t i = 1; i < mAtlasRobot->getNumBodyNodes(); ++i)
   {
     Joint* joint = mAtlasRobot->getJoint(i);
     if (joint->getDof() > 0)
     {
-      for (int j = 0; j < joint->getDof(); ++j)
+      for (size_t j = 0; j < joint->getDof(); ++j)
         joint->setDampingCoefficient(j, 80.0);
     }
   }
